requirements.txt
google-generativeai
pydantic
python-dotenv
openai
anthropic
